requirements.txt
Flask==2.1.2
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
graphene==2.1.9
graphene-sqlalchemy-filter==1.13.0
python-dateutil==2.8.2
psycopg2-binary==2.9.3
